aggregateBets helper function added
